+++ lib/polly_general_func_lib/pollyxt_display_att_beta.py
@@ -9,6 +9,7 @@ import scipy.io as spio
 import numpy as np
 from datetime import datetime, timedelta
 import matplotlib
+import python_colormap
 
 # generating figure without X server
 plt.switch_backend('Agg')
@@ -74,22 +75,6 @@ def rmext(filename):
     file, _ = os.path.splitext(filename)
     return file
 
-
-def make_colormap(seq):
-    """Return a LinearSegmentedColormap
-    seq: a sequence of floats and RGB-tuples. The floats should be increasing
-    and in the interval (0,1).
-    """
-    seq = [(None,) * 3, 0.0] + list(seq) + [1.0, (None,) * 3]
-    cdict = {'red': [], 'green': [], 'blue': []}
-    for i, item in enumerate(seq):
-        if isinstance(item, float):
-            r1, g1, b1 = seq[i - 1]
-            r2, g2, b2 = seq[i + 1]
-            cdict['red'].append([item, r1, r2])
-            cdict['green'].append([item, g1, g2])
-            cdict['blue'].append([item, b1, b2])
-    return matplotlib.colors.LinearSegmentedColormap('CustomMap', cdict)
 
 
 def pollyxt_display_att_beta(tmpFile, saveFolder):
@@ -171,18 +156,19 @@ def pollyxt_display_att_beta(tmpFile, saveFolder):
     #cmap.set_bad('k', alpha=1)
     #cmap.set_over('w', alpha=1)
     #cmap.set_under('k', alpha=1)
-    c = matplotlib.colors.ColorConverter().to_rgb 
-    cmap=make_colormap([c('lightskyblue'),c('dodgerblue'), 0.1,
-                          c('dodgerblue'),c('teal'), 0.2,
-                          c('teal'),c('limegreen'), 0.3,
-                          c('limegreen'),c('yellow'), 0.4,
-                          c('yellow'), c('orange'), 0.5,
-                          c('orange'),c('orangered'), 0.6,
-                          c('orangered'), c('red'), 0.7,
-                          c('red'), c('firebrick'), 0.8,
-                          c('firebrick'),c('darkred'), 0.9,
-                          c('darkred'),c('k')])
-    cmap.set_bad(color='white', alpha=1)
+    # c = matplotlib.colors.ColorConverter().to_rgb 
+    # cmap=eleni_colormap([c('lightskyblue'),c('dodgerblue'), 0.1,
+                          # c('dodgerblue'),c('teal'), 0.2,
+                          # c('teal'),c('limegreen'), 0.3,
+                          # c('limegreen'),c('yellow'), 0.4,
+                          # c('yellow'), c('orange'), 0.5,
+                          # c('orange'),c('orangered'), 0.6,
+                          # c('orangered'), c('red'), 0.7,
+                          # c('red'), c('firebrick'), 0.8,
+                          # c('firebrick'),c('darkred'), 0.9,
+                          # c('darkred'),c('k')])
+    # cmap.set_bad(color='white', alpha=1)
+    cmap=python_colormap.eleni_colormap
 
     # display attenuate backscatter at 355 FR
     fig = plt.figure(figsize=[10, 5])
