+++ lib/python_colormap.py
@@ -184,6 +184,37 @@ def Test():
     plt.close()
 
 
+def make_colormap(seq):
+    """Return a LinearSegmentedColormap
+    seq: a sequence of floats and RGB-tuples. The floats should be increasing
+    and in the interval (0,1).
+    """
+    seq = [(None,) * 3, 0.0] + list(seq) + [1.0, (None,) * 3]
+    cdict = {'red': [], 'green': [], 'blue': []}
+    for i, item in enumerate(seq):
+        if isinstance(item, float):
+            r1, g1, b1 = seq[i - 1]
+            r2, g2, b2 = seq[i + 1]
+            cdict['red'].append([item, r1, r2])
+            cdict['green'].append([item, g1, g2])
+            cdict['blue'].append([item, b1, b2])
+    return matplotlib.colors.LinearSegmentedColormap('CustomMap', cdict)
+
+def eleni_colormap():
+    c = matplotlib.colors.ColorConverter().to_rgb 
+    cmap=eleni_colormap([c('lightskyblue'),c('dodgerblue'), 0.1,
+                          c('dodgerblue'),c('teal'), 0.2,
+                          c('teal'),c('limegreen'), 0.3,
+                          c('limegreen'),c('yellow'), 0.4,
+                          c('yellow'), c('orange'), 0.5,
+                          c('orange'),c('orangered'), 0.6,
+                          c('orangered'), c('red'), 0.7,
+                          c('red'), c('firebrick'), 0.8,
+                          c('firebrick'),c('darkred'), 0.9,
+                          c('darkred'),c('k')])
+    cmap.set_bad(color='white', alpha=1)
+    return cmap
+
 def main():
     Test()
